test: パフォーマンステストを実施し結果をレポートにまとめる (#51)
負荷・ストレス・耐久・スパイクの4種とフロントエンド計測を実施した。
全テストでエラーは1件も出ていない。

主要な発見:

- likes / comment_likes に user_id を先頭とする索引が無く、is_liked の判定が
  毎回テーブル全件を走査している。フィード取得の実行時間の90%がこれ1つ。
  データ量に比例して悪化し、5倍にしたとき応答は25.6倍になった。
  テーブルが大きくなると PostgreSQL が並列スキャンに切り替わり、
  1リクエストあたりのCPU消費が3倍になることが超線形悪化の原因。
  索引2本の追加で p95 394.20ms -> 4.93ms (80倍) を負荷テストで実測した
- HTTPレスポンス圧縮が無効。コメント502件で 163.2KB を無圧縮で返しており、
  gzip なら 8.1KB (95.1%削減)。設定の不在は grep できないため
  実際にレスポンスを見るまで気づけなかった
- 1画面の表示に6本のAPIを呼んでいる。QueryClient に staleTime が無く
  /auth/me が2回走り、表示していないタブのフィードも取得している

事前にコードから疑っていた4点は、実測ではいずれもボトルネックでは
なかった(相関サブクエリ5本 / comment_count の二重ネスト /
LIMITなし全件取得 / 索引なしILIKE)。測る前に直していれば
効果の無い変更に時間を使っていた。

計測基盤の変更:

- PERF_SLEEP を追加。待ち時間1秒固定では 1VU=最大1req/s となり、
  VU300でも飽和に到達できなかったため
- saturate プロファイルと smoke-preauth プロファイルを追加
- perf/report/summarize.js を追加。レポートの数値を手で書き写すと
  写し間違いに気づけないため、summary JSON から機械生成する
- seed.sql の整数オーバーフローを修正(中データで初めて顕在化)
- collect.ps1 を UTF-8 BOM 付きに変更(PowerShell 5.1 が読めない)。
  max_query_sec がアイドル接続を拾う不具合も修正

.claude/skills/perf-test/ として手順をスキル化した。
実際に踏んだ罠(k6のcookie jarリセット、検証時のgrepでlevel=errorを
除外して壊れているのにOKと誤判定した件など)も併せて記録している。

計測用スタックはボリュームごと破棄済み。開発用の sns-application_pgdata が
計測前と同一であることをボリューム一覧の差分で確認した。

+++ perf/seed/seed.sql
@@ -57,7 +57,7 @@ SELECT
     t.ts,
     CASE WHEN g.n % 40 = 0 THEN t.ts + interval '1 hour' ELSE NULL END
 FROM generate_series(0, :posts_n - 1) AS g(n)
-JOIN perf_users pu ON pu.idx = (g.n * 7919) % (SELECT count(*) FROM perf_users)
+JOIN perf_users pu ON pu.idx = (g.n::bigint * 7919) % (SELECT count(*) FROM perf_users)
 CROSS JOIN LATERAL (
     SELECT now() - (g.n % 90) * interval '1 day' - (g.n % 1440) * interval '1 minute'
 ) AS t(ts);
@@ -82,8 +82,8 @@ SELECT
     p.created_at + (g.n % 240) * interval '1 minute',
     CASE WHEN g.n % 33 = 0 THEN p.created_at + interval '2 hours' ELSE NULL END
 FROM generate_series(0, (:comments_n * 4 / 5) - 1) AS g(n)
-JOIN perf_posts p  ON p.idx  = (g.n * 7919)   % (SELECT count(*) FROM perf_posts)
-JOIN perf_users pu ON pu.idx = (g.n * 104729) % (SELECT count(*) FROM perf_users);
+JOIN perf_posts p  ON p.idx  = (g.n::bigint * 7919)   % (SELECT count(*) FROM perf_posts)
+JOIN perf_users pu ON pu.idx = (g.n::bigint * 104729) % (SELECT count(*) FROM perf_users);
 
 CREATE TEMP TABLE perf_root_comments AS
 SELECT id, post_id, created_at, (row_number() OVER (ORDER BY id) - 1)::bigint AS idx
@@ -106,8 +106,8 @@ SELECT
     c.created_at + (1 + g.n % 120) * interval '1 minute',
     c.created_at + (1 + g.n % 120) * interval '1 minute'
 FROM generate_series(0, (:comments_n / 5) - 1) AS g(n)
-JOIN perf_root_comments c ON c.idx  = (g.n * 7919)   % (SELECT count(*) FROM perf_root_comments)
-JOIN perf_users pu        ON pu.idx = (g.n * 104729) % (SELECT count(*) FROM perf_users);
+JOIN perf_root_comments c ON c.idx  = (g.n::bigint * 7919)   % (SELECT count(*) FROM perf_root_comments)
+JOIN perf_users pu        ON pu.idx = (g.n::bigint * 104729) % (SELECT count(*) FROM perf_users);
 
 -- ---------------------------------------------------------------------------
 -- 5. コメントが極端に多い投稿を3件つくる
@@ -126,7 +126,7 @@ SELECT
     h.created_at + (g.n % 500) * interval '1 minute'
 FROM perf_hot_posts h
 CROSS JOIN generate_series(0, 499) AS g(n)
-JOIN perf_users pu ON pu.idx = (g.n * 104729) % (SELECT count(*) FROM perf_users);
+JOIN perf_users pu ON pu.idx = (g.n::bigint * 104729) % (SELECT count(*) FROM perf_users);
 
 -- ---------------------------------------------------------------------------
 -- 6. いいね(投稿)
@@ -140,7 +140,7 @@ SELECT
     p.created_at + (g.n % 600) * interval '1 minute'
 FROM generate_series(0, :likes_n - 1) AS g(n)
 JOIN perf_posts p  ON p.idx  = g.n % (SELECT count(*) FROM perf_posts)
-JOIN perf_users pu ON pu.idx = (g.n / (SELECT count(*) FROM perf_posts) * 7919 + g.n * 13)
+JOIN perf_users pu ON pu.idx = (g.n / (SELECT count(*) FROM perf_posts) * 7919 + g.n::bigint * 13)
                                % (SELECT count(*) FROM perf_users)
 ON CONFLICT DO NOTHING;
 
@@ -154,7 +154,7 @@ SELECT
     c.created_at + (g.n % 600) * interval '1 minute'
 FROM generate_series(0, :comment_likes_n - 1) AS g(n)
 JOIN perf_root_comments c ON c.idx  = g.n % (SELECT count(*) FROM perf_root_comments)
-JOIN perf_users pu        ON pu.idx = (g.n / (SELECT count(*) FROM perf_root_comments) * 7919 + g.n * 13)
+JOIN perf_users pu        ON pu.idx = (g.n / (SELECT count(*) FROM perf_root_comments) * 7919 + g.n::bigint * 13)
                                       % (SELECT count(*) FROM perf_users)
 ON CONFLICT DO NOTHING;
